(1749) Factorisation de la fonction reduce sur les permissions
